# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tibbiy Ko'rik Platformasi API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  super_admin: "super_admin",
  admin: "admin",
  staff: "staff",
  doctor: "doctor",
} as const;

export interface User {
  id: number;
  name: string;
  username: string;
  email?: string;
  role: UserRole;
  isActive: boolean;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  message?: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  super_admin: "super_admin",
  admin: "admin",
  staff: "staff",
  doctor: "doctor",
} as const;

export interface CreateUserRequest {
  name: string;
  username: string;
  email?: string;
  password: string;
  role: CreateUserRequestRole;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  super_admin: "super_admin",
  admin: "admin",
  staff: "staff",
  doctor: "doctor",
} as const;

export interface UpdateUserRequest {
  name?: string;
  username?: string;
  email?: string;
  password?: string;
  role?: UpdateUserRequestRole;
  isActive?: boolean;
}

export interface UsersListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export type PatientStatus = (typeof PatientStatus)[keyof typeof PatientStatus];

export const PatientStatus = {
  draft: "draft",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Patient {
  id: number;
  fullName: string;
  passport: string;
  jshshir: string;
  birthDate: string;
  phone: string;
  photoUrl?: string;
  status: PatientStatus;
  createdById: number;
  createdAt: string;
  updatedAt: string;
}

export type PatientWithRecordStatus =
  (typeof PatientWithRecordStatus)[keyof typeof PatientWithRecordStatus];

export const PatientWithRecordStatus = {
  draft: "draft",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type MedicalRecordStatus =
  (typeof MedicalRecordStatus)[keyof typeof MedicalRecordStatus];

export const MedicalRecordStatus = {
  draft: "draft",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type MedicalRecordDecision =
  (typeof MedicalRecordDecision)[keyof typeof MedicalRecordDecision];

export const MedicalRecordDecision = {
  allowed: "allowed",
  not_allowed: "not_allowed",
} as const;

export interface MedicalRecord {
  id: number;
  patientId: number;
  status: MedicalRecordStatus;
  doctorId?: number;
  chairmanId?: number;
  checkDate?: string;
  nextCheckDate?: string;
  decision?: MedicalRecordDecision;
  qrCode?: string;
  verifyCode?: string;
  uuid?: string;
  rejectionReason?: string;
  doctorName?: string;
  chairmanName?: string;
  createdAt: string;
  updatedAt: string;
}

export interface PatientWithRecord {
  id: number;
  fullName: string;
  passport: string;
  jshshir: string;
  birthDate: string;
  phone: string;
  photoUrl?: string;
  status: PatientWithRecordStatus;
  createdById: number;
  createdAt: string;
  updatedAt: string;
  medicalRecord?: MedicalRecord;
}

export interface CreatePatientRequest {
  fullName: string;
  passport: string;
  jshshir: string;
  birthDate: string;
  phone: string;
  photoUrl?: string;
}

export interface UpdatePatientRequest {
  fullName?: string;
  passport?: string;
  jshshir?: string;
  birthDate?: string;
  phone?: string;
  photoUrl?: string;
}

export interface PatientsListResponse {
  patients: Patient[];
  total: number;
  page: number;
  limit: number;
}

export type CreateMedicalRecordRequestDecision =
  (typeof CreateMedicalRecordRequestDecision)[keyof typeof CreateMedicalRecordRequestDecision];

export const CreateMedicalRecordRequestDecision = {
  allowed: "allowed",
  not_allowed: "not_allowed",
} as const;

export interface CreateMedicalRecordRequest {
  patientId: number;
  doctorId?: number;
  chairmanId?: number;
  checkDate?: string;
  nextCheckDate?: string;
  decision?: CreateMedicalRecordRequestDecision;
}

export type UpdateMedicalRecordRequestDecision =
  (typeof UpdateMedicalRecordRequestDecision)[keyof typeof UpdateMedicalRecordRequestDecision];

export const UpdateMedicalRecordRequestDecision = {
  allowed: "allowed",
  not_allowed: "not_allowed",
} as const;

export interface UpdateMedicalRecordRequest {
  doctorId?: number;
  chairmanId?: number;
  checkDate?: string;
  nextCheckDate?: string;
  decision?: UpdateMedicalRecordRequestDecision;
}

export type ApproveRecordRequestDecision =
  (typeof ApproveRecordRequestDecision)[keyof typeof ApproveRecordRequestDecision];

export const ApproveRecordRequestDecision = {
  allowed: "allowed",
  not_allowed: "not_allowed",
} as const;

export interface ApproveRecordRequest {
  doctorId?: number;
  chairmanId?: number;
  checkDate?: string;
  nextCheckDate?: string;
  decision: ApproveRecordRequestDecision;
}

export interface RejectRecordRequest {
  reason: string;
}

export interface PdfResponse {
  pdfUrl: string;
  uuid: string;
}

export interface VerifyInfoResponse {
  uuid: string;
  found: boolean;
}

export interface VerifyCodeRequest {
  code: string;
}

export interface VerifyResult {
  valid: boolean;
  patient?: Patient;
  medicalRecord?: MedicalRecord;
  message?: string;
}

export interface DashboardStats {
  totalPatients: number;
  todayChecks: number;
  approved: number;
  rejected: number;
  pending: number;
  draft: number;
  activeStaff: number;
  activeDoctors: number;
}

export interface RecentActivityResponse {
  patients: Patient[];
}

export interface DoctorsListResponse {
  doctors: User[];
}

export type GetUsersParams = {
  role?: GetUsersRole;
  search?: string;
  page?: number;
  limit?: number;
};

export type GetUsersRole = (typeof GetUsersRole)[keyof typeof GetUsersRole];

export const GetUsersRole = {
  super_admin: "super_admin",
  admin: "admin",
  staff: "staff",
  doctor: "doctor",
} as const;

export type GetPatientsParams = {
  search?: string;
  status?: GetPatientsStatus;
  page?: number;
  limit?: number;
};

export type GetPatientsStatus =
  (typeof GetPatientsStatus)[keyof typeof GetPatientsStatus];

export const GetPatientsStatus = {
  draft: "draft",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;