# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tibbiy Ko'rik Platformasi API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /medical-records:
    post:
      operationId: createMedicalRecord
      tags: [medical-records]
      summary: Create or update medical record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicalRecordRequest"
      responses:
        "200":
          description: Medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medical-records/{id}:
    get:
      operationId: getMedicalRecord
      tags: [medical-records]
      summary: Get medical record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
    put:
      operationId: updateMedicalRecord
      tags: [medical-records]
      summary: Update medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicalRecordRequest"
      responses:
        "200":
          description: Updated medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medical-records/{id}/approve:
    post:
      operationId: approveMedicalRecord
      tags: [medical-records]
      summary: Approve medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRecordRequest"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medical-records/{id}/reject:
    post:
      operationId: rejectMedicalRecord
      tags: [medical-records]
      summary: Reject medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRecordRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medical-records/{id}/pdf:
    get:
      operationId: generatePdf
      tags: [medical-records]
      summary: Generate PDF for medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: PDF URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfResponse"
...
  /verify/{uuid}/check:
    post:
      operationId: verifyCode
      tags: [verify]
      summary: Verify secret code and return full document info
      parameters:
        - name: uuid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeRequest"
      responses:
        "200":
          description: Verified document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResult"
        "401":
          description: Wrong code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user]
    User:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [super_admin, admin, staff, doctor]
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, username, role, isActive, createdAt]
...
    Patient:
      type: object
      properties:
        id:
          type: number
        fullName:
          type: string
        passport:
          type: string
        jshshir:
          type: string
        birthDate:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        status:
          type: string
          enum: [draft, pending, approved, rejected]
        createdById:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, fullName, passport, jshshir, birthDate, phone, status, createdById, createdAt, updatedAt]
...
    MedicalRecord:
      type: object
      properties:
        id:
          type: number
        patientId:
          type: number
        status:
          type: string
          enum: [draft, pending, approved, rejected]
        doctorId:
          type: number
        chairmanId:
          type: number
        checkDate:
          type: string
        nextCheckDate:
          type: string
        decision:
          type: string
          enum: [allowed, not_allowed]
        qrCode:
          type: string
        verifyCode:
          type: string
        uuid:
          type: string
        rejectionReason:
          type: string
        doctorName:
          type: string
        chairmanName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, patientId, status, createdAt, updatedAt]
    CreateMedicalRecordRequest:
      type: object
      properties:
        patientId:
          type: number
        doctorId:
          type: number
        chairmanId:
          type: number
        checkDate:
          type: string
        nextCheckDate:
          type: string
        decision:
          type: string
          enum: [allowed, not_allowed]
      required: [patientId]
    UpdateMedicalRecordRequest:
      type: object
      properties:
        doctorId:
          type: number
        chairmanId:
          type: number
        checkDate:
          type: string
        nextCheckDate:
          type: string
        decision:
          type: string
          enum: [allowed, not_allowed]
    ApproveRecordRequest:
      type: object
      properties:
        doctorId:
          type: number
        chairmanId:
          type: number
        checkDate:
          type: string
        nextCheckDate:
          type: string
        decision:
          type: string
          enum: [allowed, not_allowed]
      required: [decision]
    RejectRecordRequest:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    PdfResponse:
      type: object
      properties:
        pdfUrl:
          type: string
        uuid:
          type: string
      required: [pdfUrl, uuid]
...
    VerifyCodeRequest:
      type: object
      properties:
        code:
          type: string
      required: [code]
    VerifyResult:
      type: object
      properties:
        valid:
          type: boolean
        patient:
          $ref: "#/components/schemas/Patient"
        medicalRecord:
          $ref: "#/components/schemas/MedicalRecord"
        message:
          type: string
      required: [valid]
...