# lib/api-zod/src/generated/types/pdfResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tibbiy Ko'rik Platformasi API
 * OpenAPI spec version: 0.1.0
 */

export interface PdfResponse {
  pdfUrl: string;
  uuid: string;
}